SOBId combo change event is not working, because on change jquery async call was defined, because of that combo values are not properly getting.
At the time of saving jv, total cr/dr should be equal. need to put validation.

Allow current period date to be entered as GL date and CodeCombinition should also be get according to the date range of current period

When we update jv detail, each time description brings some kachra in it.

Account will not be repeated in GLLines

Edit not working on GLLines grid

Allow only number in Account value

Change function not implemented for Period and Currency in JV

Out of Debit and Credit, only one should be entered.